Move createUserAndJoinParty out of global.ts
